Replace small type classes with an enum Each enumeration only needs to implement isApplicable(ItemStack), which is usually a pretty simple implementation so it isnt really an issue. This also means that you do not need to go and edit TypeParameter#types every time you decide to implement a new type; it is all handled with Types.values().
